Remove factory for pooled datasource.
  * DROID was creating a new pooled datasource for each class that wanted a data source.
  * Remove the factory (we don't want to create a new data source pool for each class).
  * Put back the initialisation methods from the factory back to the data source (as it was
    originally)
  * Modify the spring.xml to instantiate the pooled data source rather than the factory.
